Problem with searching a quest's state is handled and a couple of upgrade scrolls are now working
Actually, all high and blessed scrolls are working now, and most of  the
other(middle and lwo)  classes are not working because they are not in
the ITEM table.Moreover, I haven't figure out how to include low and
middle class upgrade scroll because there is weird situation in the
structure of our ITEM table.

As you can see, I chanced QuestV2SearchEligibleQuest in such a way that
it will search the quest state according to just given sEventDataIndex.

About the database script, they musn't be in this repo, as will explain
why when we talk later.