Added/Fixed alot of more IDS Tags
Our mission now is fixing nd adding the right IDS Tags.
We are doing it because we removed the windows app which were used this
strcture IDS_XX_XX string and pull it from the client and this because
we want the client to work as well on Linux & Mac , and while in your
code your dependce on Windows app you can't achieve this goal.