GetPrivateProfileInt dependency is removed since it is windows dependent.Also
Also I'm currently trying to focus on client in starting.

As side note, we need to add
Enable=1
under [Sound] in option.ini in data folder.

Unfortunately, I'm working with VS2017 community with on windows 10, so I had to upgrade the solution to v141, so to be sure, please make sure the code is still compatible with VS2015 before merging.