Fixed broken sky issue. Resolves #59
The actual problem here was when we were setting a texture to the billboard with colors using the SetTextureState. The whole idea of the billboard, is to create an illusion of fog, so in order to let the vertices get drawn by the DrawPrimitiveUP, directx is expecting us to use the SetTexture function and we simply set it to null as we don’t need one. The SetFVF diffuse macro in combination with xyz, tells it to spread the colors around the xyz coordinates of the vertices, since each billboard(in this case front and bottom) has 4 corners.