Implement new /fog [value 1-10000] command for gms.
It would be useful for debugging or rendering or gms of course. Also notifies user when mistyped.
Set the boolean to true in GameProcMain::Init() on line 281 if you want to use it as we don't want to have weird behavior in a situation where someone forgot to add this in the Texts_us.tbl.