Command List window and Edit added (#100)
* Adding Curtail UI Class

Began adding in the command list UI.

* Fixed Closing/Opening of Dialog

Forgot to add the tick function. This fixes the opening and closing of the dialog box.

* Changed name to CUICmdList and made progress on loading cmds.

I've made some progress on loading in command categories and commands.

* Using map to load resources

I changed up the command list UI to load in the cmds through iterating over a map instead of manually.  Works fine except the fact that I get a couple gibberish commands under the private category stored in map because they are in the resources table...

8012	¾É±â/¼­±â
8013	°È±â/¶Ù±â
8014	À§Ä¡

* Updated map

Now I only load the strings and ids of the commands because the tooltip is only an offset.

* Merging changes

* Added CUICmdEdit and worked on some cmd handling

Class CUICmdEdit added and currently a little hacked to work with the in the recvmsg method to test functionality with the command list.