Login server: Auto-detect which VERSION table is used by the DB
In common usage, there's 2 sets of column names used for this table.

The official names are likely only the long form names, as they're the ones found in the old official SQL dump,
but since both are so commonly used in the wild it's worth just allowing for both of them.