Remove file format version overrides for effects
Effects are almost perfectly backwards compatible with the exception of
some extremely old effects that use a part count of 16 before they switched
it to 26, without ever bothering to allocate a version for it - which might
have even happened before any official version was released.

Aside from that, everything is versioned and entirely backwards compatible,
so we don't need to do this.