Load original string data, support changing the encoding
Ideally the list shouldn't copy this data at all; it should rely on the stored backing.
For now, we'll simply allow for it to keep the original 'ANSI' data ready to re-encode
in the editor via a menu without having to reload the file.

We also just assume Korean off the bat.
It's not always true, but it's sufficient for all but 1 file (help_us.tbl) in 1.298;
the encoding for this can just be manually changed via the menu.