Skill recast map fix (#169)
* Skill recast timer fix

It does the trick....
Seems to work better than what was supplied with the source code.
The only bug I've came across is not being able to Z or select a mob for 1-2 seconds after you cast a skill but I'm not quite sure where this happens might be in gameproc or in here.
I've only tested mage skills not the rest neither potions scrolls or items, but they should work too.
Not sure what legacy code should remain or can be removed.

* Update MagicSkillMng.cpp

* Current state of cooldown

Iterating and erasing elapsed cooldowns, map sizes decreases when cooldown does elapse double checked it under debug. Should we skip the whole computing if the list/std map is 0?