Allow for Github Desktop users as well
Github Desktop no longer installs Git alongside it, so we cannot rely on this to exist anymore.
It intends the user to install Git for Windows separately, but we'll ignore that and find its install directory.

The idea is to find its latest app directory (e.g. app-3.4.20) with git.exe inside it (in case of lingering folders
after an uninstall), determining if it's a 32-bit or 64-bit build and using that path for the update.

We continue to prefer to use git.exe as found in PATH however, but this helps it to just work out of the box.