btn behavior for character delete changed (#209)
* btn behavior for character delete changed

* style changed to mb_ok

* Preserve original behaviour and note change for reference

It would be unclear looking at this otherwise why this was changed,
since it otherwise appears like a standard yes/no prompt.

Since the resource was officially changed in 1.298 however, this
doesn't behave as you would expect, simply looking at it.

So for future reference, we'll acknowledge the reason for its change,
and preserve its original behaviour.

---------

Co-authored-by: twostars <i.am.twostars@gmail.com>