Add build workflows (#236)
* Add build workflows:

  * Add build.yml for basic client & server build (manually, on PR to master & on push to master)

  * Add build_tools.yml for the various tools (manually, and weekly every Sunday at midnight)

The client & server build is already time consuming at ~3min+, so we'll keep the tool builds
(which don't currently all build anyway) separate.

It may not necessarily be worth keeping the server build either, since most of the changes
will rely on the client.

We may want to use labels as triggers to denote which actually needs building, although when
multiple do, doing it this way will cause it to use the maximum amount of time rather than
sharing the existing libs.

* Include merge_group